add resource edit page and modify method at resources.js
