+++ pages/api/resources.js
@@ -8,14 +8,18 @@ export default async function(req, res) {
     return res.send(data);
   }
 
-  if(req.method === "POST") {
-    const { title, description, link, timeToFinish, priority } = req.body;
+  if(req.method === "POST" || req.method === "PATCH") {
+    const { id, title, description, link, timeToFinish, priority } = req.body;
     if(!title || !description || !link || !timeToFinish || !priority) {
       return res.status(422).send("Data are missing!");
     }
 
+    const url = req.method === "POST"
+      ? "http://localhost:3001/api/resources"
+      :  `http://localhost:3001/api/resources/${id}`
+
     try {
-      const axiosRes = axios.post("http://localhost:3001/api/resources", req.body);
+      const axiosRes = axios[req.method.toLowerCase()](url, req.body);
       return res.send(axiosRes.data)
     } catch {
       return res.status(422).send("Data cannot be stored!");
