+++ pages/resources/[id]/edit.js
@@ -1,0 +1,39 @@
+import Layout from "components/Layout";
+import ResourceForm from "components/ResourceForm";
+import axios from "axios";
+
+const ResourceEdit = ({ resource }) => {
+  const updateResouce = (formData) => {
+    axios.patch("/api/resources", formData)
+      .then(_ => alert("data has been updated"))
+      .catch(err => alert(err?.response?.data))
+  }
+
+  return (
+    <Layout>
+      <div className="container">
+        <div className="columns">
+          <div className="column is-8 is-offset-2">
+            <ResourceForm
+              initialData={resource}
+              onFormSubmit={updateResouce}
+            />
+          </div>
+        </div>
+      </div>
+    </Layout>
+  )
+}
+
+export async function getServerSideProps({params}) {
+  const dataRes = await fetch(`http://localhost:3001/api/resources/${params.id}`);
+  const data = await dataRes.json();
+
+  return {
+    props: {
+      resource: data
+    },
+  }
+}
+
+export default ResourceEdit;
