+++ pages/resources/[id]/index.js
@@ -1,8 +1,7 @@
+import Link from "next/link";
 import Layout from "components/Layout";
-import { useRouter } from "next/router";
 
 const ResourceDetail = ({resource}) => {
-  const router = useRouter();
 
   return (
     <Layout>
@@ -16,6 +15,11 @@ const ResourceDetail = ({resource}) => {
                       <h2 className="subtitle is-4">{resource.createdAt}</h2>
                       <h1 className="title">{resource.title}</h1>
                       <p>{resource.description}</p>
+                      <Link href={`/resources/${resource.id}/edit`}>
+                        <a className="button is-warning">
+                          Update
+                        </a>
+                      </Link>
                     </div>
                   </div>
                 </div>
